Extract utility methods into separate files
